back/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
django-ninja==1.6.2
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
soupsieve==2.8.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
